add ability to override token.properties without rebuild image

+++ docker/direct-app/build/scripts/build.sh
@@ -16,7 +16,12 @@ cp -f /root/config/run.conf /data/jboss-4.2.3.GA/bin/
 cp -f /root/scripts/start.sh /data/jboss-4.2.3.GA/bin/
 
 # copy properties files to mounted direct volume (/root/direct)
-cp -f /root/config/token.properties /root/direct/token.properties
+# change use .docker by default and fallback to in-image property file
+if [ -f /root/direct/token.properties.docker ]; then
+    cp -f /root/direct/token.properties.docker /root/direct/token.properties
+else
+    cp -f /root/config/token.properties /root/direct/token.properties
+fi
 cp -f /root/config/topcoder_global.properties /root/direct/topcoder_global.properties
 
 # build and deploy
